# server/doc/openapi.yaml
openapi: 3.0.3
info:
  title: psbFinances API
  version: 1.0.0
  description: >
    HTTP API for personal and small business finance management. All authenticated
    endpoints accept either a `Bearer` JWT in the `Authorization` header or the
    `psbf` session cookie issued by the authentication routes.
servers:
  - url: https://psbfinances.com/api
    description: Production
  - url: http://localhost:9000/api
    description: Local development
...
paths:
...
  /accounts:
    get:
      tags: [Accounts]
      summary: List financial accounts
      operationId: listAccounts
      parameters:
        - in: query
          name: include
          schema:
            type: string
          description: >
            Comma separated extras (`balance`, `unreconciled`). Both values are
            always returned when present.
      responses:
        '200':
          description: Collection of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - bearerAuth: []
        - cookieAuth: []
    post:
      tags: [Accounts]
      summary: Create an account
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountInput'
      responses:
        '200':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - bearerAuth: []
        - cookieAuth: []
  /accounts/{accountId}:
    get:
      tags: [Accounts]
      summary: Retrieve an account
      operationId: getAccount
      parameters:
        - $ref: '#/components/parameters/AccountId'
        - in: query
          name: include
          schema:
            type: string
          description: >
            Comma separated extras (`balance`, `unreconciled`). Both values are
            always returned when present.
        - in: query
          name: currentBalance
          schema:
            type: number
            format: double
          description: >
            When supplied, calculates an opening balance that reconciles with the
            provided current balance.
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - bearerAuth: []
        - cookieAuth: []
    put:
      tags: [Accounts]
      summary: Update an account
      operationId: updateAccount
      parameters:
        - $ref: '#/components/parameters/AccountId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountInput'
      responses:
        '200':
          description: Account updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
        - bearerAuth: []
        - cookieAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: psbf
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  parameters:
    AccountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
      description: Financial account identifier
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        title:
          type: string
        invalidParams:
          type: object
          additionalProperties:
            type: string
      additionalProperties: false
...
    Account:
      type: object
      properties:
        id:
          type: string
        shortName:
          type: string
        fullName:
          type: string
        type:
          type: string
          description: Account type (checking, savings, credit, etc.)
        openingBalance:
          type: number
          format: double
        currentBalance:
          type: number
          format: double
        isDefault:
          type: boolean
        closed:
          type: boolean
        visible:
          type: boolean
        deleted:
          type: boolean
        meta:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
          nullable: true
        unreconciledTransactions:
          type: integer
          nullable: true
      required: [id, shortName, fullName, type, openingBalance, visible]
    AccountInput:
      type: object
      required: [shortName, fullName, type]
      properties:
        shortName:
          type: string
        fullName:
          type: string
        type:
          type: string
        openingBalance:
          type: number
          format: double
          default: 0
        isDefault:
          type: boolean
          default: false
        closed:
          type: boolean
          default: false
        visible:
          type: boolean
          default: true
        meta:
          type: object
          additionalProperties: true
...